Moving into worlds folder
