+++ src/cpp/mcglib/CMakeLists.txt
@@ -10,16 +10,13 @@ find_package(nlohmann_json REQUIRED)
 
 file(GLOB_RECURSE mcglib_headers CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/*.h)
 file(GLOB_RECURSE mcglib_core CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/core/*.cpp)
-file(GLOB_RECURSE mcglib_zombieworld CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/ZombieWorld/*.cpp)
-file(GLOB_RECURSE mcglib_gridworld CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/GridWorld/*.cpp)
-file(GLOB_RECURSE mcglib_dungeonworld CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/DungeonWorld/*.cpp)
+file(GLOB_RECURSE mcglib_worlds CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/worlds/*.cpp)
+
 add_library(
   mcglib STATIC
   ${mcglib_headers}
   ${mcglib_core}
-  ${mcglib_zombieworld}
-  ${mcglib_gridworld}
-  ${mcglib_dungeonworld}
+  ${mcglib_worlds}
   )
 
 target_link_libraries(mcglib ${Boost_LIBRARIES} nlohmann_json::nlohmann_json)
