+++ src/cpp/mcglib/worlds/ZombieWorld/aabbs/ZombieworldPit.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../../core/AABB.h"
+#include "../../../core/AABB.h"
 
 /**
  * @brief This class represents a special AABB of type = "pit". It is not hollow
