+++ src/cpp/mcglib/worlds/ZombieWorld/blocks/Lever.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../../core/Block.h"
+#include "../../../core/Block.h"
 
 class Lever : public Block {
   private:
